Feature test services2

+++ src/main/java/innowise/khorsun/carorderservice/service/PaymentService.java
@@ -4,13 +4,12 @@ import innowise.khorsun.carorderservice.dto.PaymentDto;
 import innowise.khorsun.carorderservice.entity.Payment;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
 
 @Service
 public interface PaymentService {
     Payment addPayment(PaymentDto paymentDto);
 
-    BigDecimal calculatePaymentAmount(Integer rentalId);
+    Long calculatePaymentAmount(Integer rentalId);
 
     boolean isSessionPaid(String sessionId);
